refactor(initializeModels): split into per-capability factory modules
Move each capability's match/case dispatch block into its own file under
src/factories/:

  upscale.py    -- buildUpscaleProcess(self)
  interpolate.py -- buildInterpolateProcess(self, outputWidth, outputHeight)
  restore.py    -- RestoreChain + buildRestoreProcess(self)
  dedup.py      -- buildDedupProcess(self)
  standalone.py -- objectDetection, autoClip, segment, depth,
                   motionBlur, stabilize (unchanged logic, moved verbatim)

src/initializeModels.py is now a thin coordinator (~70 lines) that
gates on self.upscale/interpolate/restore/dedup, emits Adobe progressState
updates, and delegates to the factories. Standalone functions and
RestoreChain are re-exported so all existing call sites remain unchanged.

No logic changes; 44/44 tests pass.
